+++ include/common/configure.h
@@ -117,11 +117,11 @@ public:
         ForceInterpreter(RHS.ForceInterpreter.load(std::memory_order_relaxed)),
         AllowAFUNIX(RHS.AllowAFUNIX.load(std::memory_order_relaxed)) {}
 
-  void setMaxMemoryPage(const uint32_t Page) noexcept {
+  void setMaxMemoryPage(const uint64_t Page) noexcept {
     MaxMemPage.store(Page, std::memory_order_relaxed);
   }
 
-  uint32_t getMaxMemoryPage() const noexcept {
+  uint64_t getMaxMemoryPage() const noexcept {
     return MaxMemPage.load(std::memory_order_relaxed);
   }
 
@@ -142,7 +142,7 @@ public:
   }
 
 private:
-  std::atomic<uint32_t> MaxMemPage = 65536;
+  std::atomic<uint64_t> MaxMemPage = 65536;
   std::atomic<bool> ForceInterpreter = false;
   std::atomic<bool> AllowAFUNIX = false;
 };
