+++ include/runtime/instance/module.h
@@ -291,7 +291,7 @@ protected:
     }
     return unsafeGetMemory(Idx);
   }
-  MemoryInstance *unsafeGetMemory(uint32_t Idx) const noexcept {
+  MemoryInstance *unsafeGetMemory(uint64_t Idx) const noexcept {
     return MemInsts[Idx];
   }
   Expect<GlobalInstance *> getGlobal(uint32_t Idx) const noexcept {
@@ -324,7 +324,7 @@ protected:
     }
     return unsafeGetData(Idx);
   }
-  DataInstance *unsafeGetData(uint32_t Idx) const noexcept {
+  DataInstance *unsafeGetData(uint64_t Idx) const noexcept {
     return DataInsts[Idx];
   }
