+++ lib/system/allocator.cpp
@@ -33,7 +33,7 @@ static inline constexpr const uint64_t k12G = UINT64_C(0x300000000);
 
 } // namespace
 
-WASMEDGE_EXPORT uint8_t *Allocator::allocate(uint32_t PageCount) noexcept {
+WASMEDGE_EXPORT uint8_t *Allocator::allocate(uint64_t PageCount) noexcept {
 #if WASMEDGE_OS_WINDOWS
   auto Reserved = reinterpret_cast<uint8_t *>(winapi::VirtualAlloc(
       nullptr, k12G, winapi::MEM_RESERVE_, winapi::PAGE_NOACCESS_));
@@ -75,8 +75,8 @@ WASMEDGE_EXPORT uint8_t *Allocator::allocate(uint32_t PageCount) noexcept {
 }
 
 WASMEDGE_EXPORT uint8_t *Allocator::resize(uint8_t *Pointer,
-                                           uint32_t OldPageCount,
-                                           uint32_t NewPageCount) noexcept {
+                                           uint64_t OldPageCount,
+                                           uint64_t NewPageCount) noexcept {
   assuming(NewPageCount > OldPageCount);
 #if WASMEDGE_OS_WINDOWS
   if (winapi::VirtualAlloc(Pointer + OldPageCount * kPageSize,
@@ -106,7 +106,7 @@ WASMEDGE_EXPORT uint8_t *Allocator::resize(uint8_t *Pointer,
 #endif
 }
 
-WASMEDGE_EXPORT void Allocator::release(uint8_t *Pointer, uint32_t) noexcept {
+WASMEDGE_EXPORT void Allocator::release(uint8_t *Pointer, uint64_t) noexcept {
 #if WASMEDGE_OS_WINDOWS
   winapi::VirtualFree(Pointer - k4G, 0, winapi::MEM_RELEASE_);
 #elif defined(HAVE_MMAP) && defined(__x86_64__) || defined(__aarch64__) ||     \
